refactor(codemode): expose OpenAPI from package root
