+++ packages/codemode/package.json
@@ -7,8 +7,7 @@
   "type": "module",
   "license": "MIT",
   "exports": {
-    ".": "./src/index.ts",
-    "./adapters/openapi": "./src/adapters/openapi/index.ts"
+    ".": "./src/index.ts"
   },
   "scripts": {
     "typecheck": "tsgo --noEmit",
